refactor(BaseBallGameService): Service가 아닌 Controller로 수정
- do-while 사용으로 불필요한 코드 제거
- Service에 View 객체를 관리하고 있다는게 서비스의 역할과 상이하다고 판단, 따라서 Controller로 수정
